backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [{name = "Abi Raja", email = "abimanyuraja@gmail.com"}]
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn>=0.25.0",
    "websockets>=14.1",
    "openai==2.16.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.2",
    "httpx>=0.28.1",
    "pre-commit>=3.6.2",
    "anthropic>=0.84.0",
    "moviepy>=1.0.3",
    "pillow>=10.3.0",
    "types-pillow>=10.2.0.20240520",
    "aiohttp>=3.9.5",
    "pydantic>=2.10",
    "google-genai>=1.16.1",
    "langfuse>=3.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pyright>=1.1.352",
    "pytest-asyncio>=0.21",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=7.4.3",
    "pyright>=1.1.352",
    "pytest-asyncio>=0.21",
]
